marketstack resources and asset
