+++ .env.example
@@ -8,3 +8,4 @@ GOOGLE_SERVICE_ACCOUNT_JSON=
 ENVIRONMENT=
 DBT_TARGET=
 PYTHONLEGACYWINDOWSSTDIO=
+MARKETSTACK_API_KEY=
