+++ macro_agents/src/macro_agents/defs/resources.py
@@ -1,5 +1,7 @@
 from .resources.motherduck import motherduck_resource
 from .resources.fred import fred_resource
+from .resources.market_stack import market_stack_resource
 
 md = motherduck_resource
 fred = fred_resource
+marketstack = market_stack_resource
