--- macro_agents/src/macro_agents/defs/resources/schwab.py
@@ -1,1 +1,0 @@
-## ToDo: Implement Schwab resource
